requirements.txt
biopython
medirect
moose-classifier
pandas
psycopg2-binary
pyarrow
SCons
SQLAlchemy
taxtastic
